Extract tracing-related work packets from scheduler
Move them to plan::tracing::gc_work.
