+++ src/plan/tracing/gc_work/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod closure;
+pub(crate) mod root;
+pub(crate) mod weakref;
